+++ lib/src/constants/constants.dart
@@ -1,0 +1,5 @@
+/// Path to knowledge of network
+const String knowledgePath = 'resources/knowledge.json';
+
+/// Path to network's structure
+const String structurePath = 'resources/structure.json';